requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.6.0
pydantic-settings==2.2.1
requests==2.32.3
python-dotenv==1.0.1
streamlit==1.29.0
pandas==2.2.2
numpy==1.26.4
sentence-transformers==2.7.0
scikit-learn==1.3.2
faiss-cpu==1.8.0
pyarrow==16.0.0
torch==2.2.2 --index-url https://download.pytorch.org/whl/cpu
matplotlib
playwright==1.49.0
jinja2==3.1.4
python-multipart==0.0.9
urllib3<2.0.0
